订单与库存的整合
1.完成订单部门通过中介者对库存的调用
2.转移Mediator类并声明为单例
3.待进行与运输部门链接
4.人性化输出
