+++ src/main/java/Management/HumanResources/Staff/Driver/EngineMan.java
@@ -1,7 +1,7 @@
 package Management.HumanResources.Staff.Driver;
 
-import Marketing.OrderCenterEntity.OrderProcessingMediator;
 import Presentation.Protocol.OutputManager;
+import Mediator.DepartmentMediator;
 import Storage.TransportationCan;
 
 import java.util.Date;
@@ -35,7 +35,7 @@ public class EngineMan extends Driver {
         );
 
         //完成訂單的交付
-        OrderProcessingMediator.getInstance().deliverOneOrder(transportationCan.getOrderId());
+       DepartmentMediator.getInstance().deliverOneOrder(transportationCan.getOrderId());
 
         OutputManager.getInstance().print(
                 "运输成功！",
